Move code to the right component
